TelegramNotificationBot/requirements.txt
homematicip>=0.9.3.3
python-telegram-bot>=10.1.0 